more unsigned -> size_t

+++ src/play/display2.cc
@@ -225,7 +225,7 @@ void Display2::drawSolid(Solid const& obj)
     if ( disp->style & 2  &&  disp->size > 0 )
     {
         bodyColor(disp, obj.signature());
-        for ( unsigned ii = 0; ii < obj.nbPoints(); ++ii )
+        for ( size_t ii = 0; ii < obj.nbPoints(); ++ii )
             drawPoint(obj.posP(ii), disp);
     }
     
@@ -235,7 +235,7 @@ void Display2::drawSolid(Solid const& obj)
         bodyColor2(disp, obj.signature());
         lineWidth(disp->width);
 #if ( DIM == 2 )
-        for ( unsigned ii = 0; ii < obj.nbPoints(); ++ii )
+        for ( size_t ii = 0; ii < obj.nbPoints(); ++ii )
         {
             if ( obj.radius(ii) > 0 )
                 gleObject(obj.posP(ii), obj.radius(ii), gleCircleB);
@@ -272,7 +272,7 @@ void Display2::drawSolid(Solid const& obj)
         lineWidth(disp->width);
         bodyColor2(disp, obj.signature());
         glBegin(GL_LINE_LOOP);
-        for ( unsigned ii = 0; ii < obj.nbPoints(); ++ii )
+        for ( size_t ii = 0; ii < obj.nbPoints(); ++ii )
             gleVertex(obj.posPoint(ii));
         glEnd();
     }
@@ -304,7 +304,7 @@ void Display2::drawSphere(Sphere const& obj)
     {
         bodyColor(disp, obj.signature());
         drawPoint(obj.posP(0), disp);
-        for ( unsigned ii = obj.nbRefPoints; ii < obj.nbPoints(); ii++ )
+        for ( size_t ii = obj.nbRefPoints; ii < obj.nbPoints(); ii++ )
             drawPoint(obj.posP(ii), disp);
     }
     
@@ -312,7 +312,7 @@ void Display2::drawSphere(Sphere const& obj)
     if ( disp->style & 8  &&  disp->perceptible )
     {
         bodyColor(disp, obj.signature());
-        for ( unsigned ii = 1; ii < obj.nbRefPoints; ii++ )
+        for ( size_t ii = 1; ii < obj.nbRefPoints; ii++ )
             drawPoint(obj.posP(ii), disp);
     }
 }
@@ -412,7 +412,7 @@ void Display2::drawOrganizer(Organizer const& obj) const
             glDisable(GL_LIGHTING);
 #else
             glBegin(GL_LINES);
-            for ( unsigned ii = 0; ii < so->nbPoints(); ++ii )
+            for ( size_t ii = 0; ii < so->nbPoints(); ++ii )
                 gleVertex(so->posPoint(ii));
             glEnd();
 #endif
